requirements.txt
altgraph==0.17.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
distlib==0.3.6
filelock==3.12.2
future==0.18.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gspread==5.10.0
idna==3.4
macholib==1.15.2
oauthlib==3.2.2
pip==23.2.1
pipenv==2023.7.11
platformdirs==3.8.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyfiglet==0.8.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==68.0.0
six==1.15.0
urllib3==1.26.16
virtualenv==20.23.1
virtualenv-clone==0.5.7
wheel==0.37.0
